+++ src/contact-picker.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 
-import { AutoCompleteModule } from '@acpaas-ui/ngx-components/forms';
+import { AutoCompleteModule } from '@acpaas-ui/ngx-forms';
 
 import { ContactPickerComponent } from './contact-picker/contact-picker.component';
 import { ContactPickerService } from './contact-picker/contact-picker.service';