+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All notable changes to this project will be documented in this file.
 "### Security" in case of vulnerabilities.
 -->
 
+- [BREAKING] Upgraded the Angular version to 8 and the ACPaas UI library to version 4.
+
 ## [3.0.0] - 2020-03-11
 
 ### Added